requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
aiosqlite==0.19.0
python-multipart==0.0.6
websockets==12.0
openai-whisper==20231117
requests==2.31.0
python-dotenv==1.0.1
werkzeug==3.0.1
